Move facets+relevanceCutoff tests to TestFacetsWithRelevanceCutoffAndSortBy
Extract InjectMaxHitsIntoFacetsGroupingTest and UpdateQueryWithAffectFacetsTest
from HybridSearcherTest into a dedicated test class.
